alsa-state: Add different asound configuration for Sound Card
Currently the SGTL and Conexant Sound Cards are suppported. There
is slightly differences in the asound configuration to make it as
the default sound card for the AVS_SDK software.

Make it explicitly that differences from the recipe allow the user
to select the appropiate asound configuration at build time.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>