avs-device-sdk: simplify recipe and use https for fetching
The avs-device-sdk recipe install the AVS SDK .so libraries onto the non
standard directory (/usr/lib), so the CHRPATH needs to be modified for finding
its dependencies. Since SDK releases are constantly adding modules as .so libs
the RPATH needs to be constantly updated with each release. To avoid that
a new directory is created that collect all the libraries, so they shared the
same RPATH. In fact this new libs directory is a collection of symbolic links
that point to the real path of each library, so the directory structure is
kept.

Additionally the https protocol is used now instead of git to avoid possible
fetching issues on networks where this port is blocked.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>