avs-tools: Update json config from by setCredentials script
Use the AlexaClientSDKConfig.json from source directly and update based
on its variables. Previously a pre-filled file (on build time) was used
but this was prone to fail if the json file is changed on a future release.
To minimize this possible failure, the AlexaClientSDKConfig.json is now
filled on runtime by the setCredentials script.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>