imx7d-pico.conf: allow kernel devicetree override
use a conditional assignation to the KERNEL_DEVICETREE variable, so only in the
case is not defined (somewhere else), it defaults to the pico_pi dtb.

This way the devicetree can be overriden by the local.conf. This is useful to
select a custom devicetree when using a particular sound card. Using it along
with the MACHINE_OVERRIDE we can differentiate between different configurations
in a clean way.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>