mx7d-pico: uboot: patch diferentiation based on machine override
Using a MACHINE OVERRIDE we can diferentiate which patches to apply.
This is useful to apply a set of patches only for a specific configuration.
We would like to differentiate betwen different sound cards.

At uboot level the patch for setting a particular dtb as the default, it can
be determined by the sound card to use, which can be selected by setting
the proper MACHINE_OVERRIDE variable.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>