imx7d-pico: linux-imx: patch diferentiation based on machine override
Using a MACHINE OVERRIDE we can diferentiate which patches to apply.
This is useful to apply a set of patches only for a specific configuration.
We would like to differentiate betwen different sound cards.

At kernel level we can introduce patches for adding a particular dtb or adding
configurations only needed by an specific sound card.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>