avs-setup-demo: specify a machine override for SGTL card
kernel, uboot and alsa-state configurations are now explicitly differentiated
by a machine override. So we need to define the specific machine override we
are building for.

Before the override was only needed for conexant but now we need it for SGTL
too. Otherwise no patch on uboot and/or kernel might apply at all.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>