avs-rc-local: enable etherent at start boot
On morty the ethernet is not enabled at boot. It needs to be
manually enabled. For Alexa we assumed Ethernet is the default
network interaface and it is enabled from the rc.local file.

Also the startImage is no longer launched from the rc.local since
now is started as systemd service.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>