avs-tools: make getAVSToken specific per soundcard directory
i.MX8M images are not based on X11, so the getAVSToken needs to be
based on curl commands instead on rely on X11VNC. So this scripts
differs depending on the MACHINE OVERRIDE used and needs to be
distributed per soundcard directory.

Also add a trap to kill the python AuthServer script in case the
user aborts (by Ctrl-C) the getAVSToken.sh script, to avoid errros
when try to re-run it due to the previous python process still
running.

The getAVSToken.sh for i.MX7D remains basically the same as before
except that is stored now per soundcard directory.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>