avs-image: delete firefox and x11vnc from image
Since the authentication on Alexa SDK 1.7 can be performed on any
browser outside the target, firefox is no longer required to be
part of the image. Similar case for x11VNC that was used to open
firefox on a remote window. Even though x11vnc not stritctly
required, is still keep but only for X11 distros.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>