avs-tools: update scripts for Pico 8M and run SampleApp
Intnroduce the runAlexaSampleApp.sh to unify the way the
SampleApp is started. This way if there is a specific way to
start the sample due to different AFE (DSPC, Synaptics, etc) or
an enabled featured the runAlexaSampleApp.sh can be overwritten
with the proper commands.

setupAVS.sh and startImage.sh are updated to reflect this change
and the setUTCTime.sh now called from the runAlexaSampleApp.sh is
updated to throw a warning in case of failure.

Also adding scripts for installing DSPC for Pico 8M.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>